+++ packages/editor-core/src/editor/ui/Appearance/Comment.tsx
@@ -113,24 +113,11 @@ const SecondaryToolbar = styled.div`
 `;
 SecondaryToolbar.displayName = 'SecondaryToolbar';
 
-export interface EditorAppearanceComponentState {
-  isExpanded: boolean;
-}
+export interface EditorAppearanceComponentState {}
 
 export default class Editor extends React.Component<EditorAppearanceComponentProps, EditorAppearanceComponentState> {
-  state: EditorAppearanceComponentState = {
-    isExpanded: false
-  };
 
   static displayName = 'CommentEditorAppearance';
-
-  componentDidUpdate () {
-    const { onExpand, editorView } = this.props;
-
-    if (onExpand && editorView && !this.state.isExpanded) {
-      this.setState({ isExpanded: true }, () => onExpand!(editorView));
-    }
-  }
 
   private flashToggle = false;